Add missing OBO ontology
There was an assumption that there should be a configuration on the OBO PURL service for ontologies that I brought up in https://github.com/OBOFoundry/OBOFoundry.github.io/pull/2091 and https://github.com/OBOFoundry/purl.obolibrary.org/issues/869. This commit is adding the missing ontology with a manually curated example.