Update MIxS (#693)
Closes #685

Pinging @cmungall and @ramonawalls since they're marked as the submitter
and contact person for this record. In the previous issue
https://github.com/biopragmatics/bioregistry/issues/151, there was
discussion whether to use the mnemonics or the identifiers for MIxS.
That discussion was never resolved, but I merged `mixs` anyway with the
mnemonic, but it was before there was a provider available. Since
@only1chunts gave a provider, it makes sense to switch it over.

Co-authored-by: Chris Hunter <6037145+only1chunts@users.noreply.github.com>