Fix issues with MIRIAM alignment (#707)
Two issues came up related to recent changes:

1. `miriam:gramene.growthstage` was renamed to `miriam:gro`. This now
conflicts with `gro` which is reserved for the Gene Regulation Ontology,
so a mismatch was curated and the metaregistry links in the
`bioregistry:gramene.growthstage` record was updated
2. `miriam:tair.name` was added, this is equivalent to
`bioregistry:araport`. This is the first situation (I think) when
Identifiers.org has added a prefix that was a duplicate of something
novel in the Bioregistry but didn't have the same prefix. I'm not sure
how often this will happen in the future so I'm not sure if this kind of
problem needs addressing in the alignment code.