Add 17 prefixes for clinical trial registries used by WHO (#864)
This pull request adds 17 new prefixes for clinical trials that are
found in the World Health Organization's [integrative
database](https://trialsearch.who.int/AdvSearch.aspx). To download the
trials in that resource, do the following:

1. Navigate to https://trialsearch.who.int/AdvSearch.aspx
2. Change the recruitment status dropdown "ALL"
3. Click the "search" button
4. Click the "Export Results to XML" button on the right side
5. Click "I agree" on the pop-up
6. Click "Export all clinical trials to xml" and be patient, it's about
500MB

A few hiccups when going through the ICTRP:

1. There are Clinical Trials Information System (CTIS) trials, but their
links all go to a search page in the EU clinical trials page. I can't
find any references to them anywhere
2. https://trialsearch.who.int/ mentions The Netherlands National Trial
Register but I didn't see anything appear in the dump referring to this.
https://english.ccmo.nl/latest/news/2022/06/24/dutch-trial-register-ntr-no-longer-available
claims that the website is gone and that it should be folded into
ClinicalTrials.gov or other WHO registries
3. Peruvian Clinical Trials Registry (REPEC) is mentioned but I didn't
see any references. I went and found it anyway