Get mappings between external registries (#902)
There's an API endpoint that lets you get all if the mappings from one
external registry to another. For example, getting all mappings from OBO
Foundry to Bioportal can be done with
https://bioregistry.io/api/metaregistry/obofoundry/mapping/bioportal.

However, the implementation of this functionality is built in to the API
route itself. This PR externalizes that implementation into the
Bioregistry's manager class so it can be reused.

Related to https://github.com/INCATools/ontology-access-kit/issues/588