Refactor converter generation code (#974)
This PR gets rid of code that focuses on lists of `curies.Record`
objects and instead works directly with `curies.Converter` objects.

Along the way, this also identified issues with the data integrity on
MIRIAM, N2T, and Prefix Commons with respect to the TAIR resources
(`tair.gene` and `tair.protein`) which all used non-specific,
overlapping URLs. Therefore, these needed to get cleaned out before
being import.

Why do this? If we work directly with converters, we can make use of the
CURIE prefix reconciliation tooling to more cleanly refactor the
Bioregistry to Converter pipeline (which is causing issues when adding
prefix casing variants in a related PR #969)