Reorganize repository structure (#1103)
This PR does the following:

1. Consolidates the external registry getters (in
`bioregistry.external`), the external registry alignment classes (in
`bioregistry.align` the data artifacts (in `bioregistry.data.external`),
and a few (3) configuration files (in `bioregistry.data`) into a single
hierarchy in `bioregistry.external`.
2. Moves the metaregistry curation sheets and the raw data from the
repositories out of the `src/` structure. They're now in the
`/exports/alignment/` and `/exports/raw/` folders, respectively. The
point of this is to reduce the size of the package that gets sent to
PyPI, related to #1100
3. minor version bump to 0.11.X series

In theory, this shouldn't affect any downstream uses, since the
`bioregistry.align` submodule isn't really for external users.