Add curations to semi-automated literature workflow (#1223)
This pull request updates the `curated_papers.tsv` file with all PubMed
papers identified till 2024-10-01 in
https://github.com/biopragmatics/bioregistry/issues/1165.

Here are some statistics about the classification of each paper based on
relevancy_type so far:

**Relevant (1) classifications: 14** 
- new_prefix: 3
- new_provider: 4
- new_publication: 2
- unclear: 2**
- existing: 3

**Irrelevant (0) classifications: 40** 
- irrelevant_other: 31
- no_website: 2
- not_identifiers_resource: 7

** 1. [39104285](https://bioregistry.io/pubmed:39104285) is a provider
for UniProt IDs but was not curated due to the variable nature of the
`uri_format`. 2. [38991851](https://bioregistry.io/pubmed:38991851) was
curated as a prefix but there was some discussion about whether it
should be a provider instead. See
https://github.com/biopragmatics/bioregistry/pull/1194. Regardless, both
of these were curated as relevant (1) which seems to be the more
important classification.

---------

Co-authored-by: Mufaddal Naguthanawala <m.nguthana@hotmail.com>
Co-authored-by: Charles Tapley Hoyt <cthoyt@gmail.com>