Add more strict typing (#1253)
Related to #1201, the Bioregistry codebase already started to get unruly
when I wasn't able to give consistent code quality reviews. Moving
towards making all files being strictly typed is the next big step
towards making it more sustainable. This is a first step towards that.
It focuses on making sure that students might touch has the highest
amount of checks on it