Fix branch reuse issue in new prefix PR creation workflow (#1258)
The GitHub Actions workflow for registering new prefixes through the
issue template is reusing the same branch for each new issue, leading to
the content of the second pull request overwriting the first. This was
first noticed in this PR:
https://github.com/biopragmatics/bioregistry/pull/1240. This occurs
because the workflow does not generate a unique branch for each new
issue. As a result, subsequent PRs are force-pushing content to the
previous PR's branch instead of creating new, independent branches for
each PR.

This PR introduces the following changes to the `new_prefix_pr.yml` to
address this issue:
- Create a step to extract the issue number from the issue URL.
- Dynamically generate a unique branch name
(`create-pull-request/patch-<issue_number>`) for each new issue.
- Modify the PR creation step to use the dynamically generated branch
name, ensuring that each PR is linked to its own unique branch.

---------

Co-authored-by: Mufaddal Naguthanawala <m.nguthana@hotmail.com>
Co-authored-by: Benjamin M. Gyori <ben.gyori@gmail.com>