Add prefix: streptomedb (#1287)
This pull request curates a new prefix `streptomedb` which provides
identifiers for natural products from Streptomyces speices.

PubMed: https://pubmed.ncbi.nlm.nih.gov/39498494/
Database: http://streptomedb.vm.uni-freiburg.de:8000/streptomedb/

A couple notes regarding this resource:
- In the URL for each entry, there is a string of digits that I have
assumed is the actual identifier. There is no explicit mention of a
"Streptome ID" for any of the entries but this seemed like the most
likely case. It's entirely possible that this is an arbitrary string of
digits but given that this appears to be from a reputable source I
thought it was worth curating and modifying later if needed.
- The contributor GitHub username is the groups' username, not the
individuals.

---------

Co-authored-by: Benjamin M. Gyori <ben.gyori@gmail.com>