Commit Predictions File with Git Hash Link, Remove Artifact Upload (#1350)
This PR refactors the GitHub Actions workflow for the paper ranking
script to ensure the predictions file is committed directly to the
repository. The following changes were made:

1. Removed Artifact Upload Step:
- The actions/upload-artifact step in the workflow is currently used to
store files generated during the workflow. These artifacts are then
available for download in the GitHub Actions interface under the
"Artifacts" section for that workflow run.
- The addition of the commit and push step below makes this step
unnecessary and was removed to avoid redundant storage.

2. Added Commit and Push Step:
- Introduced a new step to commit and push the predictions file directly
to the `exports/analyses/paper_ranking/` directory in the repository.
- This ensures that the predictions file is saved in the repository for
future use if necessary.

Testing via a forked repository successfully showed that the predictions
file was saved to the correct directory once the workflow run was
complete.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Benjamin M. Gyori <ben.gyori@gmail.com>