Add test for duplicate PubMed in literature curation sheet (#1369)
I noticed `bioregistry lint` was causing some rows to get deleted, then
noticed it was because there were duplicate curations. This PR adds a
test for duplicate curations and in the five instances where there were
duplicates, I deleted the later curated row. None of these affected the
outcome, but there were conflicts in the calls from the curators.

@bgyori were you interested in doing inter-curator agreement statistics
later? if not, then we can safely delete these duplicate curations

---------

Co-authored-by: Ben Gyori <ben.gyori@gmail.com>