Avoid returning windows-style line returns (#1370)
There are few examples of data that have windows-style line returns.
This isn't a full fix - the best way would be to update all of the
sources to work on this, but is an intermediate that makes sure that the
getter methods on the Resource class at least process them out.