Add prefix: dome (#1374)
This pull request curates a new prefix 'dome', which is a set of papers
related to machine learning related publications. Machine learning has
lots of applications in life sciences, which is why I decided it would
be relevant to the bioregistry. This PR updates the json and tsv files.

This PR is the updated version of #1373 but with correct branching. 

https://pubmed.ncbi.nlm.nih.gov/39661723/

---------

Co-authored-by: Ben Gyori <ben.gyori@gmail.com>