Adjust the dates for paper_ranking pipeline to be from the first to the last of the previous month (#1402)
This pull request modifies the `start_date` and `end_date` variables of
the `paper_ranking` pipeline to be set to the first and last date of the
month respectively.

The old variables were simply subtracting 30 days from the start date
which was causing it to miss out on ranking papers from the 1st of the
month in some cases.

I also modified the `cron` job to run at midnight EST instead of
midnight UTC which is what it was set to previously.

@cthoyt