Show provenance for annotations in web app (#1095)
This is a proof of concept showing that we can propagate information
about _where_ a field displayed on the Bioregistry came from. This is
important because the Bioregistry imports data from other registries, so
it's useful to know where an annotation (e.g., the name associated with
a prefix) came from, and under what license it's being reused.