Require bioregistry prefixes conform to W3C NCNAME convention (#1449)
The W3C defines a technical standard for
[`NCNAME`](https://www.w3.org/TR/1999/REC-xml-names-19990114/#NT-NCName),
which for most intents and purposes is the same as CURIE prefixes. It
follows this form:

```
prefix := NCName
NCName := (Letter | '_') (NCNameChar)*
NCNameChar	::=	Letter | Digit | '.' | '-' | '_'
```

This PR does the following:

1. Adds a data integrity test to make sure all Bioregistry prefixes
conform to the W3C NCNAME standard
2. Adds a pattern to the Bioregistry record for itself matching the
NCNAME definition
3. Renames the three Bioregistry preixes that aren't conforming, while
maintaining the previous prefixes as synonyms:
   - `3dmet` (dead resource)
   - `4dn.replicate`
   - `4dn.biosource`
4. Adds documentation to the contribution guidelines on how to
"normalize" prefixes coming from external registries that aren't
NCNAME-conformant.
5. Updates the automated ingestion workflow to apply the underscore rule
described in 4)


This PR is a first step towards addressing
https://github.com/biopragmatics/bioregistry/issues/1338