Improve and finalize mapping checking pipeline for benchmarking (#1456)
This PR builds and improves on #1439 as follows:
- It extends mapping checking to mappings appearing in mismatch.json
which is crucial for making benchmarking unbiased while being easy to
ignore when reviewing mappings due to a boolean flag in the output tsv
to allow simply skipping these.
- It canonicalizes the metadata fields being used to `name`,
`description` and `homepage` making this consistent across the two
entries being compared. In the initial implementation this was not
carefully thought out and made consistent.
- The output tsv is extended to contain both the Bioregistry entry
metadata text used and the external registry text used for better
reproducibility of similarity scores (possible just based on the content
of the file), easier review, and consistency (in #1439 only the external
text was recorded).
- Add README explanation of columns for the output tsv.

Not reflected in this PR is that I curated a stratified random sample of
100 mappings, combined with the already curated known mismatches and got
AUROC=0.937 and AUPRC=0.947 for the similarity scores vs
correct/incorrect mappings. Some of these curations (the mismatches) can
ultimately just go into mismatches.json but the _correct_ mappings that
came up during curation would require figuring out where to put a new
file so I'm not adding any of this in this PR.