Replace mismatch file with more general mappings table (#1458)
**Background**: After #1439 and #1456 it makes sense to systematically
review certain mappings and record incorrect ones (currenty in
mismatch.json) so they can be removed. However, as part of review, one
naturally encounters correct mappings as well which need to be recorded
somewhere to enable benchmarking and to avoid re-review of the same
mappings redundantly. A further shortcoming of the current mismatch.json
is that it is structurally limited to representing a single prefix per
external registry.

**Proposal**: This PR adds a new TSV file at data/curated_mappings.tsv
which serves as a place to add the result of mapping review. I
propagated all the existing mismatches from mismatch.json into
curated_mappings.tsv. The PR also adds a documentation guide to help
with the mapping review process.

**Questions**:
- If we have a dedicated file for mapping review, what should we do with
mismatch.json?
- Option 1: just remove mismatch.json and use curated_mappings.tsv
instead as the single source of truth. I don't think there is any
downstream usage of mismatch.json so making this change shouldn't be
disruptive. Internal code related to the use of mismatch.json would have
to be updated.
- Option 2: Preserve mismatch.json but only as output of an automated
workflow that generates it from the TSV.
- Option 3: The two files could be manually kept in sync which requires
no further coding but is not great otherwise.
- Should curated_mappings.tsv actually be a TSV similar to
curated_papers.tsv or should it be JSON similar to mismatch.json just
with an enhanced data model?
- I intentionally described this as mapping "review" rather than mapping
"curation" since the idea is to review mappings that _already exist_.
Would it make sense to use this same file to curate _additional_
mappings that were not otherwise produced and use these during
alignment?

**TODOs**:
- Add validity checking for curated_mappings.tsv similar to those for
curated_papers.tsv
- Extend the curation guide: further guidance and a specific example

---------

Co-authored-by: Charles Tapley Hoyt <cthoyt@gmail.com>