Add prefix: cncb.gwh (#1452)
This PR adds a new prefix called `gwh` (The Genome Warehouse). The link
to the pubmed article and database are below:

https://pubmed.ncbi.nlm.nih.gov/39977364/
https://ngdc.cncb.ac.cn/gwh/

This repository is separated into three categories: Assembly,
Coronaviridae, and Poxviridae. This prefix is for the "Assembly" part of
the database. Each of the other categories are providers for existing
prefixes.

---------

Co-authored-by: Ben Gyori <ben.gyori@gmail.com>