Add DALIA OER and community prefix (#1616)
This pull request adds two new prefixes to the Bioregistry:

1. `dalia.oer`, which represents open educational resources in DALIA,
like
https://search.dalia.education/items/69aebdfe-73f6-4bac-85d7-047dd8cec076
for _PIDINST White Paper (chap. 4, 5)_
2. `dalia.community`, which represents communities in DALIA, like
https://search.dalia.education/communities/bead62a8-c3c2-46d6-9eb1-ffeaba38d5bf
for NFDI4Chem

After merging this PR and redeploying the Bioregistry web application,
it will be able to support making redirects like:

1. https://bioregistry.io/dalia.oer:69aebdfe-73f6-4bac-85d7-047dd8cec076
redirects to the correct OER page
2.
https://bioregistry.io/dalia.community:bead62a8-c3c2-46d6-9eb1-ffeaba38d5bf
redirects to the correct community page

> [!WARNING]  
> These links will get a 404 before this PR is merged! This is just a
demonstration of how it will look later.

The Bioregistry is not itself a PID provider like DOI (with partners
CrossRef/DataCite/etc.), but it is able to keep track of _any_ kind of
identifier scheme, and support generating more stable links.

For example, if the DALIA web frontend changes from using the URI format
`https://search.dalia.education/items/$1` for OERs to
`https://search.dalia.education/oer/$1`, the Bioregistry can reflect
this, and Bioregistry links will all continue to work.

You can see a demo of how this works for PubMed - the index of
biomedical documents.

1. The Bioregistry has a page describing PubMed at
https://bioregistry.io/registry/pubmed
2. For a given PubMed identifier, like `36888965`, you can construct a
Bioregistry link like https://bioregistry.io/pubmed:36888965. The
Bioregistry knows how to redirect to the correct page on the PubMed
website, and if it ever changes (or if PubMed gets redeployed in another
country), then we can update the Bioregistry's configuration, but the
person writing the link does not have to change anything.