Add Virtuoso prefix map validation (#1691)
Closes #1688

This can be run to validate the NFDI4Culture's endpoint

```console
$ bioregistry validate virtuoso https://nfdi4culture.de/sparql
```

note that this doesn't work for _all_ triple stores, just Virtuoso ones
because of the way that it has a special page for showing the prefix
map. Unfortunately, I couldn't figure out if there were a programmatic
way to get this (e.g., with content negotiation), so it requires doing
some HTML parsing.