Reimplement EPM construction (#1737)
This PR reimplements EPM construction by reusing upstreamed
`curies.Converter` builder functionality. It's much faster now, too.

Along the way, I had to make a few more fixes to external registry
importers and a few more tweaks to the "part of" relation to not use it
when they don't share a URI for resolving