Fill the NFDI4BIOIMAGE collection (#1791)
This PR updates the recently created NFDI4BIOIMAGE collection with the
resources relevant for that consortium.

The list of resources is largely derived from the work previously done
in the [FoundingGIDE project](https://founding-gide.eurobioimaging.eu),
which surveyed the controlled vocabularies used in the bioimaging
domain.

Related to https://github.com/nfdi-de/section-metadata-wg-onto/issues/32